requirements.txt
selenium==4.15.2
mysql-connector-python==8.2.0
openpyxl==3.1.2 